# Optimize VectorBuilder#++=(xs: TraversableOnce)