--- a/src/library/scala/collection/immutable/Vector.scala
+++ b/src/library/scala/collection/immutable/Vector.scala
@@ -655,7 +655,10 @@ final class VectorBuilder[A]() extends ReusableBuilder[A, Vector[A]] with Vector
     this
   }
 
-  override def ++=(xs: TraversableOnce[A]): this.type = super.++=(xs)
+  override def ++=(xs: TraversableOnce[A]): this.type = {
+    xs.foreach(+=(_))
+    this
+  }
 
   def result: Vector[A] = {
     val size = blockIndex + lo