We can improve the performance by using `xs.foreach(+=(_))` instead of using `super.++=(xs)`

Benchmark results are as follows:

before
```
[info] Benchmark                      (size)  Mode  Cnt          Score        Error  Units
[info] VectorBenchmark.builderConcat       1  avgt   20      61304.834 ±   5649.373  ns/op
[info] VectorBenchmark.builderConcat      10  avgt   20     143109.259 ±  22050.573  ns/op
[info] VectorBenchmark.builderConcat     100  avgt   20    1087507.944 ±  15564.617  ns/op
[info] VectorBenchmark.builderConcat    1000  avgt   20    9975918.443 ±  52197.275  ns/op
[info] VectorBenchmark.builderConcat   10000  avgt   20  101155017.712 ± 988804.970  ns/op
```

after
```
[info] Benchmark                      (size)  Mode  Cnt         Score         Error  Units
[info] VectorBenchmark.builderConcat       1  avgt   20     38958.564 ±    1012.915  ns/op
[info] VectorBenchmark.builderConcat      10  avgt   20    116479.188 ±     916.548  ns/op
[info] VectorBenchmark.builderConcat     100  avgt   20    789017.215 ±    4176.121  ns/op
[info] VectorBenchmark.builderConcat    1000  avgt   20   7428021.120 ±   41471.026  ns/op
[info] VectorBenchmark.builderConcat   10000  avgt   20  74379030.621 ± 1083102.872  ns/op
```