@Ichoran 2017-02-24
Can you fix this in `Growable` instead?  The problem is that the special loop case is `LinearSeq` when it should be `LinearSeqOptimized`.