@monkey-mas 2017-02-25
Thanks for your feedback! I might be wrong but it seems like I can't fix `Grawable#++=(...)` as I want to use [VectorBuilder#+=(...)](https://github.com/scala/scala/pull/5740/files#diff-59f3462485b74027de4fd5e9febcc81bL646) to optimize the code.