@monkey-mas 2017-02-25
> The problem is that the special loop case is LinearSeq when it should be LinearSeqOptimized

Can you please be more specific? What do I exactly have to do for `Growable#++=(...)` in this case?