@Ichoran 2017-02-25
Just change `LinearSeq` to `LinearSeqOptimized` and benchmark again.  `Vector` is a `LinearSeq`, but it has a relatively slow head/tail decomposition.  It is _not_ a `LinearSeqOptimized`, so it will fall back to the `foreach(b += _)` branch.