@Ichoran 2017-02-27
It extends both IndexedSeq, because it can index more efficiently than O(n), and LinearSeq because it can do a head/tail decomposition more efficiently than O(n). It does _not_ extends LinearSeqOptimized because head/tail decomposition is not among the fastest operations it can do.