@xuwei-k 2017-02-27
```scala
$ scala -language:_
Welcome to Scala 2.12.1 (Java HotSpot(TM) 64-Bit Server VM, Java 1.8.0_112).
Type in expressions for evaluation. Or try :help.

scala> :paste
// Entering paste mode (ctrl-D to finish)

def getSuperClasses(clazz: Class[_]): List[Class[_]] = {
    @scala.annotation.tailrec
    def loop(o: Class[_], result: List[Class[_]]): List[Class[_]] = {
      val superClass = o.getSuperclass
      if (superClass == null) {
        result
      } else {
        loop(superClass, superClass :: result)
      }
    }
    loop(clazz, List(clazz))
  }

  def allSuperClassAndTrait(clazz: Class[_]): List[Class[_]] = {
    def loop(c: Class[_], result: List[Class[_]]): List[Class[_]] = {
      val interfaces = c.getInterfaces.toList
      if (interfaces.size == 0) {
        c :: result
      } else {
        interfaces.flatMap{ i => loop(i, i :: result) }
      }
    }

    val classes = for {
      c <- getSuperClasses(clazz)
      result <- loop(c, Nil)
    } yield result

    (classes.toSet - clazz).toList.sortBy(_.toString)
  }


// Exiting paste mode, now interpreting.

getSuperClasses: (clazz: Class[_])List[Class[_]]
allSuperClassAndTrait: (clazz: Class[_])List[Class[_]]

scala> allSuperClassAndTrait(classOf[Vector[_]]).foreach(println)
class java.lang.Object
interface java.io.Serializable
interface scala.Equals
interface scala.Function1
interface scala.Immutable
interface scala.PartialFunction
interface scala.Serializable
interface scala.collection.CustomParallelizable
interface scala.collection.GenIterable
interface scala.collection.GenIterableLike
interface scala.collection.GenSeq
interface scala.collection.GenSeqLike
interface scala.collection.GenTraversable
interface scala.collection.GenTraversableLike
interface scala.collection.GenTraversableOnce
interface scala.collection.IndexedSeq
interface scala.collection.IndexedSeqLike
interface scala.collection.Iterable
interface scala.collection.IterableLike
interface scala.collection.Parallelizable
interface scala.collection.Seq
interface scala.collection.SeqLike
interface scala.collection.Traversable
interface scala.collection.TraversableLike
interface scala.collection.TraversableOnce
interface scala.collection.generic.FilterMonadic
interface scala.collection.generic.GenericTraversableTemplate
interface scala.collection.generic.HasNewBuilder
interface scala.collection.immutable.IndexedSeq
interface scala.collection.immutable.Iterable
interface scala.collection.immutable.Seq
interface scala.collection.immutable.Traversable
interface scala.collection.immutable.VectorPointer
```

> It extends both IndexedSeq, because it can index more efficiently than O(n), and LinearSeq 

I don't think so. Vector extends only IndexedSeq in Scala 2.12.1