@Ichoran 2017-02-27
@xuwei-k - Oh!  You're right.  My mistake!

Sorry, @monkey-mas!

Still, can you test the speed of adding a list?  That's what the special-case in `Growable` is for.  Vector-vector additions are not test enough of a speed improvement.

I really hate to have to copy superclass code to subclasses to get decent performance.  The algorithm isn't different at all.