@lrytz 2017-03-27
Moving to M2 as this is still being discussed and doesn't change / add any functionality