@szeiger 2017-07-04
@monkey-mas Do you have time to address the remaining comment? We're planning to close the queue for M2 until the end of next week.